Complete rewrap of the code to be fairly autonomous wrt scikit-learn
I was depending on sklearn for the coding (OMP, LARS etc) and on the
learning (with the scripts in dict_learning). Due to the tim it takes to
get my PR through, I decided to keep some methods of sparse coding in
the scripts (I notably call LARS, etc..) but
1- to create my own matching pursuit function (approx. 10 lines)
2- to incorporate most of my PR in this code.

As a consequence, the code is with the defaults parameters independent
of sklearn, but may occasionally use it for comparing other sparse
coding mechanisms.
As a side effect, it is likely that some wording  / formatting of the
sklearn library was used... I have had a hard time diffing etc to sort
out my part from that of the community. At least, it runs autonomously!