Implements a non-parametric rescaling function
I was not happy with the C parameter in the rescaling function which had
to be changed from experiment to experiment.
This does not change the unsupervised learning but may lead to numerical
errors. So instead of using an exponential as the rescaling function, I
use instead a vector computed from the sorted coefficients. It seems to
work sufficiently good. To use it, set C to zero.

Test and dev notebook @ http://blog.invibe.net/posts/2017-11-07-meul-with-a-non-parametric-homeostasis.html